% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{counter00X}
\alias{counter00X}
\title{Counter}
\usage{
counter00X(counter0)
}
\arguments{
\item{counter0}{value to be advanced by 1.}
}
\value{
counter
}
\description{
This funciton advances a value of 'counter0' +1 each time it is used.
}
\examples{
counter00X(1)
"002"
}
