% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{EditCommData}
\alias{EditCommData}
\title{Takes a string of words and combines them into a sentance that lists them.}
\usage{
EditCommData(dat, category0)
}
\arguments{
\item{dat}{data from the NOAA Fisheries One Stop Shop https://foss.nmfs.noaa.gov/apexfoss/f?p=215:200::::::}

\item{category0}{string. Name of a column that contains the category names.}
}
\description{
This function alows you to automatically edit data into the correct format from the NOAA Fisheries One Stop Shop https://foss.nmfs.noaa.gov/apexfoss/f?p=215:200:::::: plus a column with how you want to categorize your data.
}
\examples{
EditCommData()
}
\keyword{Commercial}
\keyword{Data}
\keyword{Edit}
