% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{itis_reclassify}
\alias{itis_reclassify}
\title{Title}
\usage{
itis_reclassify(tsn, categories, missing.name)
}
\arguments{
\item{tsn}{A vector of Taxonomic Serial Numbers to be evaluated.}

\item{categories}{A list of the categories and associated TSN values.}

\item{missing.name}{A string of what to call the missing value.}
}
\value{

}
\description{
Title
}
\examples{

}
