% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{lmCheck}
\alias{lmCheck}
\title{Linear Regression Check}
\usage{
lmCheck(Columns, temp)
}
\arguments{
\item{Columns}{Names of columns to apply this action to.}

\item{temp}{Name of dataset to apply this action to.}
}
\description{
This function provides a general mental check - how many of your individual data are significantly increaseing, decreasing, or not changing? This will help the user understand the results they get from their analysis.
}
\examples{
lmCheck()
}
