% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{numbers0}
\alias{numbers0}
\title{How Many Speices are in a Dataset numeric identifier}
\usage{
numbers0(x)
}
\arguments{
\item{x}{x is a string of all the numbers you are interested in 'standardizing'.}
}
\description{
This funciton standardizes the length of the category or species numbers e.g.,(numbers of 33, 440, and 1 are converted to 033, 440, and 001)
}
\examples{
numbers0(x = c(1,14,302))
}
