% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{OutputAnalysis}
\alias{OutputAnalysis}
\title{Run Analysis for the US and several regions.}
\usage{
OutputAnalysis(
  landings_data,
  category0,
  baseyr,
  titleadd,
  dir_analyses,
  reg_order = c("National", "North Pacific", "Pacific", "Western Pacific (Hawai`i)",
    "New England", "Mid-Atlantic", "Northeast", "South Atlantic", "Gulf of Mexico"),
  reg_order_abbrv = c("US", "NP", "Pac", "WP", "NE", "MA", "NorE", "SA", "GOM"),
  skipplots = FALSE,
  save_outputs_to_file = TRUE
)
}
\arguments{
\item{landings_data}{Landings data with the following columns: "Year", "Pounds", "Dollars", category0, "Tsn", "State"}

\item{category0}{A character string. The column where the category is defined.}

\item{baseyr}{Numeric year (YYYY). The base year you are assessing the anaylsis with. Typically this is the earliest year in the data set, but it can be any year you choose.}

\item{titleadd}{A string to add to the file with the outputs to remind you why this particular analysis was interesting.}

\item{dir_analyses}{A directory that your analyses will be saved to (e.g., "./output/").}

\item{reg_order}{The US and each region that you would like to assess. Default = c("National", "North Pacific", "Pacific", "Western Pacific (Hawai`i)", "New England", "Mid-Atlantic", "Northeast", "South Atlantic", "Gulf of Mexico").}

\item{reg_order_abbrv}{Acronym of the US and each region listed in reg_order. Default = c("US", "NP", "Pac", "WP", "NE", "MA", "NorE", "SA", "GOM").}

\item{skipplots}{TRUE (create and save plots) or don't FALSE.}

\item{save_outputs_to_file}{TRUE (save outputs from analysis within function) or don't FALSE.}
}
\value{
warnings_list, editeddata_list, index_list, spp_list, figures_list, gridfigures_list
}
\description{
Run Analysis for the US and several regions.
}
\examples{
vignette('FEUS-tables')
}
