% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{PriceChange}
\alias{PriceChange}
\title{Caluclate the price change.}
\usage{
PriceChange(R0, P0)
}
\arguments{
\item{R0}{Names of columns with revenue share data.}

\item{P0}{Names of columns with price or price index data.}
}
\description{
This function allows you to caluclate the price change.
}
\examples{
PriceChange()
}
\keyword{Change,}
\keyword{Index,}
\keyword{Price}
\keyword{Price,}
\keyword{Revenue}
\keyword{Revenue-share}
\keyword{share,}
