% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{PriceIndex}
\alias{PriceIndex}
\title{Caluclate the price Index}
\usage{
PriceIndex(dat, BaseColName, baseyr, var = "PC")
}
\arguments{
\item{dat}{Default data.}

\item{BaseColName}{The base names (the leading category name) of columns with price or price index data.}

\item{baseyr}{Numeric year (YYYY). The base year you are assessing the anaylsis with. Typically this is the earliest year in the data set, but it can be any year you choose.}

\item{var}{Vairable you want to find the price index from. Default = "PC". These characters preceed your data column name 'BaseColName'}
}
\description{
This function allows you to caluclate the price change.
}
\examples{
PriceIndex()
}
\keyword{Change,}
\keyword{Index,}
\keyword{Price}
\keyword{Price,}
\keyword{Revenue}
\keyword{Revenue-share}
\keyword{share,}
