% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{PriceMethodOutput}
\alias{PriceMethodOutput}
\title{Price Method}
\usage{
PriceMethodOutput(temp, baseyr, title0 = "", place = "", category0)
}
\arguments{
\item{temp}{Default dataset.}

\item{baseyr}{Numeric year (YYYY). The base year you are assessing the anaylsis with. Typically this is the earliest year in the data set, but it can be any year you choose.}

\item{title0}{Title of analysis}

\item{place}{Area you are assessing the analysis for. This can also be used as a title.}

\item{pctmiss}{Percent missing threshold. For 60 percent use "0.6." Default is 1.00.}

\item{MinimumNumberOfSpecies}{An integer indicating the minimum number of species the user is willing to use in an analysis. If set to 1, the analysis will run even if the category only has one species. If set to 10, the analysis will run if there is 10 , 11, or more species, but not 9 or less species. Default = 2.}
}
\description{
This function calculates the Implicit Quanity Output at Fishery Level by systematically runing the Price Method Productivity Output analysis for all species of each cateorgy.
}
\examples{
PriceMethodOutput()
}
