% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{PriceMethodOutput_Category}
\alias{PriceMethodOutput_Category}
\title{Price Methods - Category Level}
\usage{
PriceMethodOutput_Category(
  temp,
  ii,
  category,
  category0,
  baseyr,
  maxyr,
  minyr,
  warnings_list = ls()
)
}
\arguments{
\item{temp}{Default dataset.}

\item{ii}{Category number.}

\item{category}{A character string. A unique string from the 'category0' column of the group being evaluated.}

\item{category0}{A character string. The column where the category is defined.}

\item{baseyr}{Numeric year (YYYY). The base year you are assessing the anaylsis with. Typically this is the earliest year in the data set, but it can be any year you choose.}

\item{maxyr}{The maxium year to assess in the dataset.}

\item{minyr}{The minium year to assess in the dataset.}

\item{warnings_list}{A list where warnings are stored. If using this function in the PriceMethodOutput it will be inherited. If using outside of that function, put ls().}
}
\description{
This function systematically runs the Price Method Productivity Output analysis for all species of a cateorgy.
}
\examples{
PriceMethodOutput_Category()
}
