% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{PriceMethodOutput_Plots}
\alias{PriceMethodOutput_Plots}
\title{Create Price Methods Plots}
\usage{
PriceMethodOutput_Plots(
  temp,
  baseyr,
  title0 = "",
  NumberOfSpecies,
  figures.list = ls()
)
}
\arguments{
\item{temp}{Default dataset.}

\item{baseyr}{Numeric year (YYYY). The base year you are assessing the anaylsis with. Typically this is the earliest year in the data set, but it can be any year you choose.}

\item{NumberOfSpecies}{A character string of the 'number of species.' Not a great name, but if you have 103, 3, 8888 species, your 'NumberOfSpecies' is "000", "0", and "0000", respectively.}

\item{figures.list}{A list where these popular figures are stored. If using this function in the PriceMethodOutput it will be inherited. If using outside of that function, put ls().}

\item{title}{Title of plot.}
}
\description{
This function produces popular plots for the Price Method Productivity Output analysis.
}
\examples{
PriceMethodOutput_Plots()
}
