% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{QuantityMethodOutput_Category}
\alias{QuantityMethodOutput_Category}
\title{Quantity Methods - Category Level}
\usage{
QuantityMethodOutput_Category(
  temp,
  ii,
  baseyr,
  maxyr,
  minyr,
  pctmiss,
  warnings.list,
  MinimumNumberOfSpecies = 1
)
}
\arguments{
\item{temp}{Default dataset.}

\item{ii}{Category number.}

\item{maxyr}{The maxium year to assess in the dataset.}

\item{minyr}{The minium year to assess in the dataset.}

\item{pctmiss}{Percent missing threshold. For 60% use "0.6." Default is 1.00.}

\item{warnings.list}{A list where warnings are stored. If using this function in the PriceMethodOutput, it function, it will be inherited. If using outside of that function, put ls().}

\item{MinimumNumberOfSpecies}{An integer indicating the minimum number of species the user is willing to use in an analysis. If set to 1, the analysis will run even if the category only has one species. If set to 10, the analysis will run if there is 10 , 11, or more species, but not 9 or less species. Default = 1.}
}
\description{
This function systematically runs the Quantity Method Productivity Output analysis for all species of a cateorgy.
}
\examples{
QuantityMethodOutput_Category()
}
