% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{ReplaceFirst}
\alias{ReplaceFirst}
\title{Replace the first value, if missing, with the next nearest value.}
\usage{
ReplaceFirst(colnames, temp)
}
\arguments{
\item{colnames}{Names of columns to apply this action to.}

\item{temp}{Name of dataset to apply this action to.}
}
\description{
If the first value of the timeseries of this column (c) is 0/NaN/NA. Change the first value (and subsequent 0/NaN/NA values) to the first available non-0/NaN/NA value. Then, used in before with 'ReplaceMid'.
}
\examples{
ReplaceFirst()
}
\keyword{First}
\keyword{Replace,}
