% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{ReplaceMid}
\alias{ReplaceMid}
\title{Replace the first value, if missing, with the next nearest value.}
\usage{
ReplaceMid(colnames, temp)
}
\arguments{
\item{colnames}{Names of columns to apply this action to.}

\item{temp}{Name of dataset to apply this action to.}
}
\description{
If a middle value of the timeseries of this column (c) is 0/NaN/NA. Change the currently 0/NaN/NA value to the previous available non-0/NaN/NA value. Then, used after with 'ReplaceFirst'.
}
\examples{
ReplaceMid()
}
\keyword{Mid,}
\keyword{Middle}
\keyword{Replace,}
