% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{TFP_ChangeRate_Method2}
\alias{TFP_ChangeRate_Method2}
\title{Another method for calculating TFP}
\usage{
TFP_ChangeRate_Method2(temp.output, temp.input, baseyr, pctmiss)
}
\arguments{
\item{temp.output}{Dataset from productivity output.}

\item{temp.input}{Dataset from productivity input.}

\item{baseyr}{Numeric year (YYYY). The base year you are assessing the anaylsis with. Typically this is the earliest year in the data set, but it can be any year you choose.}

\item{pctmiss}{Percent missing threshold. For 60 percent use "0.6." Default is 1.00.}
}
\description{
Another method for calculating TFP
}
\examples{
TFP_ChangeRate_Method1()
}
