% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{xunits}
\alias{xunits}
\title{Standardize Units}
\usage{
xunits(val, combine = T)
}
\arguments{
\item{val}{Value to be evaluated.}

\item{combine}{TRUE/FALSE (Default = TRUE). Asks if you want two strings (FALSE) or 1 concatenated string (TRUE).}
}
\description{
This funciton standardizes units of a value. For example, 1,000,000 would become "1 Million."
}
\examples{
xunits(1234567890)
"1.2 billion"
}
