% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{itis_reclassify}
\alias{itis_reclassify}
\title{Reclassify ITIS species based off a list of higher taxonomic groupings}
\usage{
itis_reclassify(tsn, categories, missing_name)
}
\arguments{
\item{tsn}{A vector of Taxonomic Serial Numbers to be evaluated.}

\item{categories}{A list of the categories and associated TSN values.}

\item{missing_name}{A string of what to call the missing value.}
}
\value{

}
\description{
Reclassify ITIS species based off a list of higher taxonomic groupings
}
\examples{
itis_reclassify(tsn = c(83677, 172746),
categories = list("Finfish" = 914179, # Infraphylum	Gnathostomata
"Shellfish" = c(82696, # Phylum	Arthropoda  – Artrópode, arthropodes, arthropods
69458)), # Phylum	Mollusca  – mollusques, molusco, molluscs, mollusks
missing_name = "uncategorized")
}
