% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{plotnlines}
\alias{plotnlines}
\title{Plot n lines in ggplot}
\usage{
plotnlines(dat, titleyaxis, title0)
}
\arguments{
\item{dat}{Default data.}

\item{titleyaxis}{y-axis title.}

\item{title0}{Title of plot.}
}
\description{
This funciton plots n lines in a ggplot.
}
\examples{
xunits(1234567890); "1.2 billion"
plotnlines()
}
