% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{PriceMethodOutput}
\alias{PriceMethodOutput}
\title{Price Method}
\usage{
PriceMethodOutput(dat00, baseyr, title0 = "", place = "", category0)
}
\arguments{
\item{dat00}{Dataset.}

\item{baseyr}{Numeric year (YYYY). The base year you are assessing the anaylsis with. Typically this is the earliest year in the data set, but it can be any year you choose.}

\item{title0}{Title of analysis}

\item{place}{Area you are assessing the analysis for. This can also be used as a title.}

\item{category0}{A character string. The column where the category is defined. A character string.}
}
\description{
This function calculates the Implicit Quanity Output at Fishery Level by systematically runing the Price Method Productivity Output analysis for all species of each cateorgy.
}
