% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{tornb}
\alias{tornb}
\title{Tornqvist Price Index Base Year Function}
\usage{
tornb(dat, Year, pvar, vvar, prodID, base.year)
}
\arguments{
\item{dat}{The dataset you would like to use.}

\item{Year}{Name of the column holding year data.}

\item{pvar}{Name of the column holding price data.}

\item{vvar}{Name of the column holding value data.}

\item{prodID}{Name of the column holding prodID data.}

\item{base.year}{The year dollar values need to be in.}
}
\description{
Tornqvist Price Index Base Year Function
}
\examples{
tornb()
}
