% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{tornc}
\alias{tornc}
\title{Tornqvist Price Index Base Year chain Function}
\usage{
tornc(dat, Year, pvar, qvar = NA, vvar, prodID, base.year)
}
\arguments{
\item{Year}{Name of the column holding year data.}

\item{pvar}{Name of the column holding price data.}

\item{qvar}{Name of the column holding quantity data.}

\item{vvar}{Name of the column holding value data.}

\item{prodID}{Name of the column holding prodID data.}

\item{base.year}{The year dollar values need to be in.}

\item{data1}{The dataset you would like to use.}
}
\description{
Tornqvist Price Index Base Year chain Function
}
\examples{
tornc()
}
