% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishEconProdOutput.R
\name{tornc}
\alias{tornc}
\title{Tornqvist Price Index Base Year chain Function}
\usage{
tornc(dat, Year = "Year", pvar = "p", vvar = "v", prodID = "prod", baseyr)
}
\arguments{
\item{dat}{The dataset you would like to use.}

\item{Year}{Name of the column holding year data.}

\item{pvar}{Name of the column holding price data.}

\item{vvar}{Name of the column holding value data.}

\item{prodID}{Name of the column holding prodID data.}

\item{baseyr}{The year dollar values need to be in.}
}
\description{
Tornqvist Price Index Base Year chain Function
}
\examples{
tornc(dat = data.frame("Year" = c(2001:2020, 2001:2020, 2001:2020, 2001:2020),
                       "p" = rnorm(n = 80, mean = 1, sd = .1),
                       "v" = rnorm(n = 80, mean = 500, sd = 300),
                       "prod" = c(rep_len("A", 20), rep_len("B", 20), rep_len("C", 20), rep_len("D", 20))),
     Year = "Year",
     pvar = "p",
     vvar = "v",
     prodID = "prod",
     baseyr = 2015)
}
